# Conflicts:
#	src/org/isf/accounting/manager/BillBrowserManager.java
#	src/org/isf/dicom/manager/DicomManagerFactory.java
#	src/org/isf/dlvrrestype/manager/DeliveryResultTypeBrowserManager.java
#	src/org/isf/dlvrrestype/service/DeliveryResultTypeIoOperation.java
#	src/org/isf/exa/manager/ExamBrowsingManager.java
#	src/org/isf/exa/manager/ExamRowBrowsingManager.java
#	src/org/isf/exa/service/ExamIoOperations.java
#	src/org/isf/exa/service/ExamRowIoOperations.java
#	src/org/isf/exatype/service/ExamTypeIoOperation.java
#	src/org/isf/hospital/service/HospitalIoOperations.java
#	src/org/isf/lab/manager/LabManager.java
#	src/org/isf/medicals/manager/MedicalBrowsingManager.java
#	src/org/isf/medicalstockward/manager/MovWardBrowserManager.java
#	src/org/isf/medstockmovtype/manager/MedicaldsrstockmovTypeBrowserManager.java
#	src/org/isf/medtype/manager/MedicalTypeBrowserManager.java
#	src/org/isf/operation/service/OperationIoOperations.java
#	src/org/isf/operation/service/OperationRowIoOperations.java
#	src/org/isf/supplier/manager/SupplierBrowserManager.java
#	src/org/isf/supplier/service/SupplierOperations.java

+++ src/org/isf/medicalstock/manager/MovStockInsertingManager.java
@@ -1,5 +1,6 @@
 package org.isf.medicalstock.manager;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.GregorianCalendar;
 import java.util.List;
@@ -118,7 +119,7 @@ public class MovStockInsertingManager {
 						OHSeverityLevel.ERROR));
 			}
 			if (GeneralData.LOTWITHCOST) {
-				Double cost = lot.getCost();
+				BigDecimal cost = lot.getCost();
 				if (cost == null || cost.doubleValue() <= 0.) {
 					errors.add(new OHExceptionMessage("zeroLotCostError",
 							MessageBundle.getMessage("angal.medicalstock.multiplecharging.zerocostsnotallowed"), //$NON-NLS-1$
